Rename run_container to system_container to reflect use of systemd
